requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.120.4
fonttools==4.57.0
h11==0.16.0
idna==3.11
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
threadpoolctl==3.5.0
typing-extensions==4.13.2
tzdata==2025.2
uvicorn==0.33.0
zipp==3.20.2
